code cleanup

+++ SwagFaq/src/Resources/app/administration/src/locations/faq-list.vue
@@ -3,36 +3,41 @@ import faqItem from '../components/faq-item.vue';
 import { data } from '@shopware-ag/meteor-admin-sdk';
 import { computed, onMounted, ref } from 'vue';
 import { MtButton, MtTextField } from '@shopware-ag/meteor-component-library';
+import EntityCollection from '@shopware-ag/meteor-admin-sdk/es/_internals/data/EntityCollection';
 
 const { Criteria } = data.Classes;
 
-const product = ref<EntitySchema.Entity<'product'>|undefined|null>(undefined);
+const product = ref<EntitySchema.Entity<'product'> | null>(null);
 const faqEntryRepository = data.repository('swag_faq_entry');
 
 const faqEntries = computed({
-    get(): EntitySchema.Entity<'swag_faq_entry'>[] {
-        return product.value?.extensions?.swagFaqEntries ?? [];
+    get(): EntitySchema.EntityCollection<'swag_faq_entry'> | null {
+        return product.value?.extensions
+            ?.swagFaqEntries as EntitySchema.EntityCollection<'swag_faq_entry'> | null;
     },
-    set(value): void {
-        console.log('setter called', value);
-        product.value.extensions.swagFaqEntries = value;
+    set(value: EntitySchema.EntityCollection<'swag_faq_entry'> | null): void {
+        if (product.value?.extensions) {
+            product.value.extensions.swagFaqEntries = value;
+        }
 
         data.update({
             id: 'sw-product-detail__product',
             data: {
                 extensions: {
-                    swagFaqEntries: value
-                }
-            }
+                    swagFaqEntries: value,
+                },
+            },
         });
-    }
+    },
 });
 
 onMounted(async () => {
-    product.value = await data.get({
+    // Todo: on page reload when on the faq tab this is sometimes null
+    // Todo: how can we wait for it to be there?
+    product.value = (await data.get({
         id: 'sw-product-detail__product',
-        selectors: ['id', 'versionId', 'extensions']
-    }) as EntitySchema.Entity<'product'>;
+        selectors: ['id', 'versionId', 'extensions'],
+    })) as EntitySchema.Entity<'product'>;
     const criteria = new Criteria();
     criteria.addFilter(
         Criteria.multi('and', [
@@ -41,31 +46,39 @@ onMounted(async () => {
         ])
     );
 
-    const entries = await faqEntryRepository.search(criteria);
-
-    faqEntries.value = entries;
+    faqEntries.value = await faqEntryRepository.search(criteria);
 });
 
 async function onDelete(id: string) {
+    if (!faqEntries.value) return;
+
     // Todo: add confirm modal
     const toDelete = faqEntries.value.find((entry) => entry.id === id);
     if (!toDelete) {
         return;
     }
     await faqEntryRepository.delete(toDelete.id);
-    faqEntries.value = faqEntries.value.filter((entry) => entry.id !== id);
+    faqEntries.value = faqEntries.value.filter(
+        (entry) => entry.id !== id
+    ) as EntityCollection<'swag_faq_entry'>;
 }
 
 const newQuestion = ref('');
 const newAnswer = ref('');
 async function onAddButtonClicked() {
     const newEntry = await faqEntryRepository.create();
-    newEntry.question = newQuestion.value;
-    newEntry.answer = newAnswer.value;
+    if (newEntry) {
+        newEntry.question = newQuestion.value;
+        newEntry.answer = newAnswer.value;
 
-    faqEntries.value = [...faqEntries.value, newEntry];
-    newQuestion.value = '';
-    newAnswer.value = '';
+        // Todo this is ugly, is there another way to trigger the setter?
+        faqEntries.value = [
+            ...(faqEntries.value ?? []),
+            newEntry,
+        ] as EntityCollection<'swag_faq_entry'>;
+        newQuestion.value = '';
+        newAnswer.value = '';
+    }
 }
 </script>
 
@@ -85,7 +98,7 @@ async function onAddButtonClicked() {
             {{ $t('swagFaq.form.add') }}
         </MtButton>
     </div>
-    <hr>
+    <hr />
     <p v-if="faqEntries?.length === 0">
         {{ $t('swagFaq.list.emptyMessage') }}
     </p>
